The prepublish hook does not work when installing deps through git and not npm.
For now, we will have to push each distribution into git.  This will remain the case to avoid having to install this packages dev deps to build the distro of this package.  I feel this compromise is worse compared to just having a committed version of each version in git.  Does require developer to run prepublish before commit.